refactor(cli): always print errors when using the cli
Prior to this change, `slick run` did not print any errors that were
caused by the underlying run implementation, which required the user to
use `go run` manually to find out what's wrong.

This change fixes that, and also pipes the stderr of all commands into
the OS' stderr which should help with debugging things.

To achieve that, I abstracted the command logic and added the
`commandOptions` type. I chose an options type over passing a `verbose`
parameter directly, this keeps things more flexible and maintainable
down the line.