fix(run): app printing invalid listen address
Prior to this change, when executing `slick run` on a newly generated,
unmodified project, the app would print the address
`http://localhost::3000` when run. Trying to quick-open this URL in the
browser results in an error/empty page due to the redundant colon
between host and port.

This change fixes that and the app now displays a browsable address when
running.

Note: There's obviously room for more sanitization etc., but I did not
want this change to exceed the intent of this app at this stage. For me,
I just want to play around with slick and these changes simply make the
experience a little more smooth :)